Provide example configuration
Provide an example configuration which is automatically made as the default configuration if no other configuration is provided. This makes starting the server for development as per the instructions in the README simpler and works without any additional actions. Without this, there is no config file and the site will not be served.