reilly edits
I want the rewards out of YouTube sync. It has no place here and confuses the purpose of YouTube sync to the reader, which is to save a publisher labor if they have hundreds of archival videos. YouTube Sync is not about rewards, and we need to stop making it part of rewards verification.