Fixed Docker script and README
Updated the Docker script to start the site correctly. The DIR variable was not being used in each place which it should have been used and this has allowed the removal of the `readlink -f` command which had compatibility issues with MacOS (#297).

The README did not have the change in name for the docker script reflected in all references so this has also been corrected.