Update mining-pools.md
We finally did it! The Luxor community has asked for a mining pool to mine LBRY, and today we are officially launching a new pool. We are happy to continue to provide our miners with more options and to support the LBRY community. For those of you who are unfamiliar with Luxor, we are a altcoin pool built for high quality cryptocurrencies. We originally started with Sia and joined the Decred community soon after.
Luxor only mines cryptos we believe in, and we’ve been keeping tabs on LBRY for the past few months. LBRY is a digital marketplace that is controlled by the market’s participants rather than a corporation or other 3rd-party. It is an open, fair, and efficient marketplace for digital goods, with an incentive design that encourages scale. Creators and publishers can charge for their content through the built in currency LBC or put up content for free.
With Suprnova controlling a majority of mining market share, it is crucial that other reliable pools enter, in order to decentralize the network. For LBRY to succeed as a business, mining must be decentralized. So join us at Luxor to help LBRY succeed.