Create id.yaml for i18n Internationalization LBRY
When I'm upset, I'm dizzy thinking about the beautiful girl I saw in the market. and confused to shoot the girl with a LOVE BULLETS so that she becomes the lover of my heart and MARRIED to be my queen and become the queen of MEMELANDIA EMPIRES.
Suddenly the INITIATIVE appeared to translate en.yaml to Indonesian and a file named id.yaml was published in our beloved LBRY environment.